New flyby enemy, enemies spawn times can now be floats, multiple enemies can now be spawned on the same update frame, made terrain bigger, reorganized prefabs folder

+++ Assets/Scripts/Enemies/EnemyBase.cs
@@ -11,7 +11,7 @@ public class EnemyBase : MonoBehaviour
     public float HitRegTimer = 0.3f;
     public float HitRegTimerReset = 0.3f;
 
-    public int SpawnTime;
+    public float SpawnTime;
 
     // Start is called before the first frame update
     void Start()
